// types/rolling/r-factor/bloom-model.ts
export interface BloomModel {
  SEMIPRODUCT_CODE: string;
  SEMIPRODUCT_ID: string;
  PROD_SIZE: string;
  PROFILE_TYPE: string;
  JOBTEXT: string;
  JOB_ID: string;
  STAND: number;
  RFACTOR: number;
  TORQUE: number;
  INCHESPERSECOND: number;
  RFACTORSETPOINT: number;
  CROSSSECTIONAREA: number;
  MESSAGEDATE: Date;
}